cliniq/Oral-Dental/oral-classify/textvision_scripts/requirements.txt
torch
torchvision
pillow
matplotlib
pandas
gdown
tqdm
